requirements_simple.txt
streamlit
python-dotenv
openai
datasets
langchain-core
langchain-openai
Pillow
